--- src/app/account/verify-2fa/page.tsx
@@ -1,5 +1,0 @@
-import { VerifyTwoFactor } from '@/app/account/verify-2fa/verify-two-factor';
-
-export default function Page() {
-    return <VerifyTwoFactor />;
-}
