+++ src/app/profile/actions/setup-two-factor-action.ts
@@ -17,7 +17,9 @@ export const setupTwoFactorAction = actionClient
         const session = await getSession();
 
         if (!session) {
-            throw new ActionError('You must be signed in to configure two-factor authentication');
+            throw new ActionError(
+                'You must be signed in to configure two-factor authentication'
+            );
         }
 
         const { base32, otpauth } = generateTOTPSecret(session.email);
